Move routing logic to a separate file
