+++ src/main/kotlin/com/ferreter/plugins/Routing.kt
@@ -1,24 +1,16 @@
 package com.ferreter.plugins
 
 import com.ferreter.client.api.AlphaVantageApi
+import com.ferreter.routes.simpleStockQuoteRoute
 import io.ktor.server.application.*
-import io.ktor.server.response.*
 import io.ktor.server.routing.*
 
 fun Application.configureRouting(
-    alphaVantageApi: AlphaVantageApi
+    alphaVantageApi: AlphaVantageApi,
 ) {
     val apiKey = environment.config.propertyOrNull(KTOR_ENV)?.getString()
     routing {
-        get("/") {
-            apiKey?.let {
-                val quote = alphaVantageApi.getQuote(
-                    symbol = "IBM",
-                    apiKey = apiKey,
-                )
-                call.respond(quote)
-            } ?: call.respond("Failed to get the API key!")
-        }
+        simpleStockQuoteRoute(apiKey, alphaVantageApi)
     }
 }
